Controller Implementation
Fixed the jump code so that you actually jump off the ground but need to add parameters so that we dont have the classic infinite jump. made the groundpoints to begin this process in hopes to check if the player is on the ground so we can jump. Added some stuff for the player to be able to use the controller to walk around and jump. The system made for the controller is completely different from that of the keyboard because simply I needed it to be like that.  Also I want a way to implement the fade shader but havent really found a good way to do that yet.